// model/aircraftType.ts
export type Aircraft = {
  id: number;
  name: string;
  type: string;
  engines: number;
  qualification: string;
  cruiseSpeedktas: number;
  rangenm: number;
  emptyWeightlbs: number;
  fuelCaplbs: number;
  maxPayloadlbs: number;
  maxTakeOfflbs: number;
  pax: number;
  cost: number;
  engineType: number;
  wheelsType: number;
};